+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "virtual-list",
-  "version": "19.1.2",
+  "version": "19.1.3",
   "author": {
     "name": "Evgenii Grebennikov",
     "email": "djonnyx@gmail.com"
