+++ src/app/app.component.html
@@ -1,7 +1,7 @@
 <div class="vl-section block cap">
   <h1>ng-virtual-list</h1>
   <p class="l">Maximum performance for extremely large lists</p>
-  <p><span>v 19.1.2</span></p>
+  <p><span>v 19.1.3</span></p>
 </div>
 
 <div class="vl-section">
